requirements/dev.txt
-r base.txt
coverage==4.2
flake8==3.2.1
isort==4.2.5
mixer==5.6.6
pytest==3.0.5
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-sugar==0.7.1
pytest-xdist==1.15.0
vcrpy==1.10.3
